A different proposal for explaining USB issues
I thought this version provides the reader with a better idea of their options. I would still like to know how to recover from accidentally removing a USB stick before detaching it from the VM w/o rebooting. I think it would be great if we could include instructions here on how to do it.