Fix qubes-os.org formatting problems
Some indentation and line spacing are mangling code blocks when displayed with the qubes-os.org stylesheet. Trying to correct...
In "Spaces" mode, I tried to get the editor to insert spaces instead of tabs, but it was converting my spaces into tabs :(
I also moved the warning about NM to the NM section.