Fixed a couple grammar / syntax issues
The "s" in the first section was out of place, so I put it on the VM instead of the work.

Also, I'm really not sure what the standard way to refer to a VM that's used as a template. Is this a "template VM" or a "TemplateVM"? I modified a few that I found, but I realized half way through that there was inconsistency all over the place. ("Standalone VM", "AppVM", etc.) Is there a standard way of writing this?