Update user-faq.md about keyboard layout
I know this feels like a workaround for some misbehaving/missing UI, but it is better than the reference to the ML discussion, and it links to the open issue.

I have not done extensive testing to this method. I am sending the PR in a hurry because I have been frustrated (missing my compose-key) for some days after reinstalling, and I think this may prove useful for other users. Please try to reproduce and/or get some community feedback before accepting the PR;  Also, feel free to reword, reformat, or discard :)