Edit pentesting.md
Updated the wikipedia definition to reflect what's on the page (which is a more accurate definition anyways). It's a little heavy, but since the sentence has quotation marks we should be basing it on a direct quote from the article and not a paraphrase. If you want to paraphrase it because it's dense someone else can come and remove those quotation marks and write their own defintion. 

While I did not modify the source, I also want to ask that we not source information from the wikipedia article itself, but rather the sources that it references in its footnotes. 

Also:
* Missing a lette
* Removed unnecessary word glyphs