Add link for resizing root disk image
This page has a bit of overlap with the Resize Root Disk Image page; the latter is a bit more clear and specific, and also describes caveats such as to avoid changes exceeding 10G between reboots. Added link to the other page, and updated some of the guidance to be more clear.