Add instructions for newer UEFI systems
Add instructions on how to create a USB installation medium that works on newer ThinkPad systems in UEFI-only mode. This is necessary because existing troubleshooting steps for ThinkPad and UEFI are no longer sufficient. The content was created based on the thread at https://groups.google.com/forum/#!topic/qubes-users/TEmVIozLJh0