Create minimal template instructions for each version
The minimal template upgrade instructions still included the
`qvm-trim-template` command, which only works in R3.2. I've moved those
instructions into the R3.2 section and added new instructions using the
`fstrim` command in the R4.0 section.

I've also removed the section on differences between the standard and
minimal upgrade procedures, since the second point is no longer accurate
(e.g., `sudo` is installed by default in the minimal template on R3.2),
and the first point is not significant enough to merit keeping the
section.

https://github.com/QubesOS/qubes-doc/pull/505
https://github.com/QubesOS/qubes-issues/issues/3429