Improve instructions for randomizing MAC addresses
NetworkManager, by default, uses a connection ID and a per-host random
and secret key to generate `stable` MAC addresses. The intention is to
keep a connection's MAC address stable indefinitely but for it to be
different on every host.

The current instruction mention that "`stable` generates a random
address that persists for each boot session". This is indeed true for
AppVMs using stock TemplateVMs. The reason is that the secret key doesn't
exist in the template and thus is only created when the AppVM starts.
This, however, may not be true for other VMs.

In order to ensure that MACs are always only `stable` during one boot
session, `stable-id`, which is used to generate MACs, can be adjusted.
NetworkManager's documentation suggests to use `${CONNECTION}/${BOOT}`
to ensure generated MACs are unique to a boot session and connection [1].

[1]: https://developer.gnome.org/NetworkManager/stable/nm-settings.html