Refactor USB document into two
This splits the usb.md document into two in preparation for
updating them with R4.0 information. It was already lengthy,
and expanding it would have made it worse. The "common tasks" of
attaching and detaching USB drives and non-mouse and keyboard
devices are left in common-tasks/usb.md. The rarely used items
such as creation/deletion of a USB qube and attaching mouse and
keyboard are being moved to configuration/usb-qube.md. Finally,
the master document list is updated to reflect this split.