Add FAQ item about coreboot
As touched on in IRC with holger and marek. New and existing users are often confused about what hardware to buy and whether coreboot is superior to vendor bios.  I've tried to keep this entry concise (it could already be considering too long) what I've tried to do is include the most often asked questions without writing a tutorial which is probably out of scope for the Qubes website and better left to the coreboot wiki. I've ignored libreboot here as it is not part of the certified hardware requirements and I don't recall a report of success.