Add secure boot question
As discussed in IRC with marek and h0lger. I've basically summarised what marek said during discussion / from old mailing list thread on same. There appear to be a lot of (to not put too find a point on it) crap discussion on the mailing list about secure boot which may mislead skim reading users. People often ask what binaries they need to sign, so I have included list, but it may be better to omit as we have not tested it (which I've also noted).