Update fedora-minimal.md
I found the existing doc misleading, as it claims sudo is not installed. This misled me to try to install sudo, which dnf said was already there. Also, the fact that the template does not have passwordless root is not a bug and therefore the relevant install should not be described as a "fix" in the doc.

I have also included explicit instructions to enable passwordless root, which will also serve as an example to newcomers of how to install the packages listed in the following section.