Fixed line numbering for new DispVM instructions
Line numbering was pushed to far to the right causing all steps to be numbered "1" not 1,2,3... as  it should. This error showed up in https://www.qubes-os.org/doc/dispvm-customization/#create-custom-sys-net-sys-firewall-and-sys-usb-dispvms . However, was not visible in master branch i.e. shows lines are numbered correctly.

I believe moving line numbering all the way to the left margin will fix the problem but I'm not 100% sure on this.