Update hvm.md
It's fairly easy to miss the fact that some of the text is for R3 and some of it is for R4. In the source it's pretty straightforward (Since it contains a linebreak) but when rendered, said linebreak is removed. Adding a few third-level headers clears this up, at least in opinion.