Document Split GPG error
Split GPG doesn't work when the private key is protected by a passphrase in the
GPG backend. It took me a couple hours to find the problem and how to fix it,
first because the ioctl error is not common and then because of the
pinentry-ncurses limitation. I'm hoping that future users that search for "qubes
gpg ioctl error" will find this information.