Reorganize mentions to passphrases in Split GPG
I added a mention about the limitation regarding passphrase-protected GPG keys
in the *Setting up the GPG backend domain* section, which is the one that users
will follow up in their initial set up. Given that this issue is not specific of any
QubesOS version I don't think it deserves a mention in those sections.