Windows 10 Qube for Dummies
The motivation to write this down was not that it is very
complicated (obviously it is not), but rather that the docs
I have read (before just trying it) somehow left a bad
impression about Windows on Qubes.

I had the impression that:

- if one wants Windows, one must use Windows 7 (which is
  EOL 1/2020, so only a short term option, if any)
- newer Windows don't work or are unstable because the
  tools needed are not tested / not working on them.

Both is not true IF one can live with a live with a very
basic featureset, which I also described.