Add separators between mailing list sections
For better visual distinction, add separators between the
mailing list sections. The visual differences between the
many headings and subheadings on the page are very small
such that overlooking the beginning of a new mailing list
section becomes likely. As a result, people could send their
emails to a wrong list by accident which already happened.