Add solution to make persistent rules in /etc/hosts
Add solution to make persistent rules in /etc/hosts. 
I checked that it works, but I hope the approach will be checked before approval for being qubes-way and clean enough.
If it's not, different solution should be provided to modify /etc/hosts permanently for TemplateBasedVMs, because, it seems, there is no one in the documentation.