Fix garbled text in sudoers file
There's a stray line in the sudoers file in the docs which isn't in the actual file in the distro. The problem was introduced by commit 658e02cc5090cfc8b7d94d18871a224d8541d3e8 which should have removed the text "incidentally by one of the Qubes developers (RW)".