Remove permalinks and update redirects
In order to support release-specific documentation
(QubesOS/qubes-issues#5308), we want to allow permalinks for
release-specific pages to be generated by _config.yml rather than
being manually specified in the YAML header of each individual page.

However, we want the unversioned URLs to point to the latest version of
each page. Currently, it seems like best (but still undesirable) way to
do this is by specifying those unversioned URLs in the redirect_from
fields in the YAML headers on each page.