Update intro.md
What I have tried to do is condense the page and remove information that I believe is a better fit for the FAQ page. If a comparison is still needed ("How does Qubes OS compare to XYZ"), I suggest a table rather than an entire section. Also, a section on : is "Qubes OS good for me" might be a good idea. Finally, a diagram will eventually be incorporated.