Update intro design and add diagram
I have tried to incorporate your remarks and to make the intro more visually appealing. There is a blend of markdown and HTML. I may need some suggestions on how to make it look ok (although, on the local web server, it looks fine) #871.

The diagram attachment needs to be added first. Also, if accepted, parts if not all of the former introduction need to be moved to the FAQ section.