Adds a current build results matrix.
The table is added to "spare a lot of building time to new users, smooth the learning curve, and save the world!". By looking at the table one can quickly understand the working/not-working build combos, and focus on his own needs instead of reinventing the wheel (and wasting his own time or, worse, add duplicate issues in the tracker).