make config file example more accurate
please confirm my changes are correct. 

if we are showing an example of applying some configuration values (like ... `= true`), then it should be not be commented out, to make clear this is what the user needs to do to apply a configuration.

also if the default of audio_low_latency is true, then (and in my dom0 it is the case) that it should be by default set to `true` not `false`. also updated the language to reflect what the `guid.conf` looks like by default.