Add screenshot and link to troubleshooting
Also capitalize the "S" in "Qube Settings". Unfortunately, this is
somewhat inconsistent between the Qube Manager and the Qubes Domains
widget. Hopefully, capitalizing both words makes it clear that this
is also the title of something, rather than the commonplace but
erroneous capitalization of the term "qube" to refer to a VM.
See QubesOS/qubes-issues#4723.

This is a follow-up to #1071.