Move "Managing AppVM Shortcuts" to Troubleshooting
As mentioned in #1071, this page is really about troubleshooting.
Reading this page normally shouldn't be required, so it doesn't
really belong under "Common Tasks."

The only exception to this would be non-Linux VMs for which a
manual sync is required. That's one argument in favor of moving
this under "Advanced Configuration" rather than "Troubleshooting,"
but at least half of this page is undeniably about troubleshooting.

There's also a "behind the scenes" section at the end that belongs
in the Developer Documentation.