Rename "awesome" to "AwesomeWM"; update image, links
Based on examining the AwesomeWM project's web presences and Wikipedia
entry, there does not appear to be a single consistent name by which the
project refers to its window manager. All three of these variations are
used: "awesome", "Awesome", and "AwesomeWM". In order to avoid ambiguity
with the regular English word, we're opting for "AwesomeWM" in our own
documentation.

This commit also reverts the change to the widget image, since that
change was not necessary (see #1161).

Finally, this commit adds links to the KDE, i3, and AwesomeWM pages.