Update minimal-templates.md
Yubikey works out of the box for 2fa responses in debian-11.template but not in the minimal. It isn't straight forward to know which package to include for that functionality to work. Others on the forum have had this issue also. I've found that package and included it here so others can easily add that functionality.