Add note to bind-dirs when dir does not exist yet
The doc on bind-dirs only provided an example where /var/lib/tor
was made persistent in sys-whonix. However, if one tries to make
persistent any directory that wasn't already present in the
template qube, it would not persist as expected. The issue
QubesOS/qubes-issues#5862 demonstrates this exact confusion.

This commit adds a note about this situation.

Fixes QubesOS/qubes-issues#5862