Remove custom install page; redirect to installation guide
This page is severely out-of-date and has remained so for an extended
period of time, with no one able or willing to update it. The outdated
content has proven misleading or unhelpful to many users. Since we have
nothing with which to replace the outdated content, removing it seems
like the lesser of two evils. (Note that the content is still available
via Git history and other archives for historical reasons and for any
users still on EOL releases.) If anyone is able and willing to
contribute the relevant knowledge and/or text for customizing the
installation of a supported Qubes release, we can bring this page back
in the future.

Closes QubesOS/qubes-issues#5632
Closes QubesOS/qubes-issues#5637
Closes QubesOS/qubes-issues#7915