Use the 4.1 policy file location and format.
The current document still refers to /etc/qubes-rpc/policy and uses
the old 3-column format. This commit updates it to use
/etc/qubes/policy.d and and 5+-column format.

Additionally, correct a technical inaccuracy in the final example
(the policy which the reader is instructed to create denies requests
from source_vm2 with the +testfile1 argument, but the prose states
that it will be accepted).