gui-troubleshooting.md: Improve the section about external screens
I was affected by this problem. When I saw this documentation, I didn't realize this was the problem affecting me. I don't have a 4K monitor; I have a wide monitor. I can click on most things, just not in specific sections of the screen.

Rewrite to be more generic.