Apply a more consistent style to introduction
Following the doc migration from Markdown to RestructuredText, a more
logical markup was applied to the *Features* section, but not to the
section about *video tours*, *screenshots* and *Getting started*. Now,
the links are better, too.

The `figure` markup should be used for images with a caption or a
description. All the images on this page are purely decorative, thus the
alt should be empty. Two images around a title gave a weird result with
the new html style so the second one was moved.