How to use devices: better cross-referencing
A lot of options of ``qvm-device`` command have been converted to Sphinx
options directives and references.

The `action-1` role doesn't seem to be used and the duplicate title was
producing a warning.

The Qubes Device Widget tray icon is presented as a figure, to get a
more consitent paragraph.

SYNOPSIS have been changed to code blocks.

The Device Widget icon is updated (from
(qubes-desktop-linux-manager)[QubesOS/qubes-desktop-linux-manager/tree/release4.2/].
As the original PNG was deleted, all the ``how-to-use-*-devices`` have
been updated too.
QubesOS/qubes-issues#4692 is closed so the corresponding paragraph has
been removed.