Use red-tinted icons and change title when PR
By checking some environment variable from Read The Docs in conf.py, we
can determine if the build is made from a PR build. In that case, the
HTML title is slightly different and instead of normal icons, we use
red-tinted ones.

That way, putting some links to the PR builds overview will be clearing
distinct from the official build.